moved core game logic to `phichain-game`
