+++ phichain-editor/src/tab/game/mod.rs
@@ -1,14 +1,12 @@
 pub mod core;
 pub mod hit_effect;
 pub mod illustration;
-pub mod scale;
 pub mod ui;
 
 use bevy::{prelude::*, render::camera::Viewport};
 
 use crate::project::project_loaded;
 use crate::tab::game::hit_effect::HitEffectPlugin;
-use crate::tab::game::scale::ScalePlugin;
 
 use self::{core::CoreGamePlugin, illustration::IllustrationPlugin, ui::GameUiPlugin};
 
@@ -21,7 +19,6 @@ impl Plugin for GameTabPlugin {
                 PostUpdate,
                 update_game_camera_viewport.run_if(project_loaded()),
             )
-            .add_plugins(ScalePlugin)
             .add_plugins(GameUiPlugin)
             .add_plugins(IllustrationPlugin)
             .add_plugins(CoreGamePlugin)
