+++ phichain-game/src/constants.rs
@@ -1,0 +1,5 @@
+use bevy::prelude::Color;
+
+// the color for perfect hit particles and lines
+// #feffa9
+pub const PERFECT_COLOR: Color = Color::rgb(254.0 / 255.0, 1.0, 169.0 / 255.0);
